refactor: create separate type for diagram node and edge and extract util function in separate file
